order_service/requirements.txt
grpcio==1.56.2
grpcio-tools==1.56.2
protobuf==4.23.3
